src/requirements.txt
ansicolors
pyinstaller
pytest
pytest-cov
pytest-mock